% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/district_query.R
\name{doogal_download}
\alias{doogal_download}
\title{Download postcodes from doogal.co.uk for a district}
\usage{
doogal_download(district)
}
\arguments{
\item{district}{name of a local authority district (lower tier only)
(England or Wales only)}
}
\value{
a data frame of postcodes and related information
}
\description{
TODO possibly enable querying the Counties API etc as well.
}
\examples{
head(doogal_download("Isles of Scilly"))
}
