% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_api.R
\name{query_api}
\alias{query_api}
\title{Get Data from Ofcom Mobile Coverage API}
\usage{
query_api(postcode)
}
\arguments{
\item{postcode}{a postcode}
}
\value{
The API response (JSON format)
}
\description{
"Subscribers can retrieve broadband coverage data against postcodes
up to 500 calls/minute for 150,000 requests per month."
}
