% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapioca-package.R
\docType{package}
\name{tapioca-package}
\alias{tapioca}
\alias{tapioca-package}
\title{tapioca: Tap into the Ofcom Mobile API}
\description{
This package facilitates querying the Ofcom Mobile API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/francisbarton/tapioca}
  \item Report bugs at \url{https://github.com/francisbarton/r-tapioca/issues}
}

}
\author{
\strong{Maintainer}: Fran Barton \email{francis.barton@citizensonline.org.uk}

}
\keyword{internal}
