% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_query.R
\name{doogal_download}
\alias{doogal_download}
\title{Download postcodes from doogal.co.uk for a district/ltla or a county/utla}
\usage{
doogal_download(area)
}
\arguments{
\item{area}{The name of a local authority district, unitary authority or county
(England or Wales only).}
}
\value{
A data frame of postcodes and related information.
}
\description{
Download postcodes from doogal.co.uk for a district/ltla or a county/utla
}
\examples{
head(doogal_download("Isles of Scilly"))
}
